src/requirements.txt
beautifulsoup4==4.13.4
openai==1.76.0
python-dotenv==1.1.0
Requests==2.32.3
schedule==1.2.2
tqdm==4.65.0
lxml==4.9.3